extracting apps grid into its own file
